Day 29 - Implementing more SSE Trigonometry
Most of today went into adapting code of [SSE Math Functions extensions]
(https://github.com/to-miz/sse_mathfun_extension/) to
my tastes. It now works, in my opinion much better than my
yesterday's attempts.

I'm done with Trig for now. I copied over [cephes]
(http://www.netlib.org/cephes/) math library and maybe even
eventually will attempt to implement `asin` and `acos` (the latter,
as it turns out, being a fancy inversion of `asin`). But I'm a bit
fatigued by now. This is too much of an unfamiliar ground for me,
so I'll try to avoid it.

I spent the rest of the day reshuffling the file structure (again), as
well as dependency structure (again). In the end, I settled for
having a  separate `math` folder, and started redefining my vector
types in the newly created `sparrow_vector_types.h`. Now the work
will begin to bring [HandmadeMath]
(https://github.com/HandmadeMath/Handmade-Math) into my project.

+++ src/console/console_platform.h
@@ -48,13 +48,13 @@ local b32 ButtonHover(struct console* Console, v2i CursorPos, v2i TopLeft, char*
 inline void
 V2iGridClamp(struct console* Console, v2i* Pos)
 {
-    if ((u32)Pos->x >= Console->Size.Width) {
+    if (Pos->x >= Console->Size.Width) {
         Pos->x = Console->Size.Width - 1;
     }
     else if (Pos->x < 0) {
         Pos->x = 0;
     }
-    if ((u32)Pos->y >= Console->Size.Height) {
+    if (Pos->y >= Console->Size.Height) {
         Pos->y = Console->Size.Height - 1;
     }
     else if (Pos->y < 0) {
@@ -64,13 +64,13 @@ V2iGridClamp(struct console* Console, v2i* Pos)
 inline void
 V2fGridClamp(struct console* Console, v2f* Pos)
 {
-    if ((u32)Pos->x >= Console->Size.Width) {
+    if (Pos->x >= Console->Size.Width) {
         Pos->x = 0;
     }
     else if (Pos->x < 0) {
         Pos->x = (f32)Console->Size.Width - 1;
     }
-    if ((u32)Pos->y >= Console->Size.Height) {
+    if (Pos->y >= Console->Size.Height) {
         Pos->y = 0;
     }
     else if (Pos->y < 0) {
